requirements.txt
streamlit>=1.32.0
streamlit-option-menu>=0.3.6
pyyaml>=6.0
gspread
oauth2client
python-dotenv